+++ theme/mode/index.ts
@@ -1,0 +1,9 @@
+import { darkTheme } from "./dark";
+import { lightTheme } from "./light";
+
+export const themes = {
+  light: lightTheme,
+  dark: darkTheme,
+};
+
+export * from "./creator";
